Outsorce home content.
This allows easy editing for the user as markdown has a way simpler
syntax than most other. Moreover, it does support plain HTML which also
allows the user to really modify it to his liking without ever touching
the system templates. However, in the future the location of this file
might need to change.